Added initial support for friends authentication.
This is a bit of a hack.

Electron doesn't support 'prompt', but it is impossible
to add it ourselves since it is a synchronous call.

So we implement it using an asynchronous call and then have the
completion of that call trigger the authentication again
when we can then return the previously entered value.

This allows us to support friends authentication at the cost
of a guaranteed failed authentication on the first attempt.